feat: upgrade MUX stream ID from 32-bit to 64-bit and add redis-mux-proxy
- Upgrade MUX protocol stream ID from uint32 to uint64 to prevent
  stream ID wraparound in long-running proxy scenarios
- Update kernel MUX implementation (mux.c/mux.h) for 64-bit stream ID
- Update hiredis MUX client library for 64-bit stream ID
- Update redis-benchmark --mux mode for 64-bit stream ID
- Implement redis-mux-proxy with full MUX multiplexing support:
  - Worker thread pool with epoll-based event loop
  - MUX backend connection management
  - Client connection handling with RESP protocol parsing
  - Hot upgrade support with zero-downtime connection migration
  - Admin server for management commands
  - Configurable via command-line options
- Add test suites: kernel MUX tests, proxy stability/concurrent/mixed tests
- Add hot upgrade persistent test script

+++ proxy/include/hot_upgrade.h
@@ -52,7 +52,7 @@ private:
         int fd;
         uint16_t listen_port;
         bool authenticated;
-        uint32_t stream_id;
+        uint64_t stream_id;
         char backend_addr[64];
         uint16_t backend_port;
         uint32_t recv_buf_len;
